+++ CHANGELOG.md
@@ -2,19 +2,27 @@
 
 # encrypt-string-plugin Changelog
 
-## [Unreleased]
+## Unreleased
+
+## 1.2.0-beta - 2024-01-25
+
 ### Added
+
 - when an error occurs during decrypt, a notification is shown in the ide.
 - Line marker on yaml file with encrypted string (with ENC(...))
 
-## [1.1.0] - 2023-07-05
+## 1.1.0 - 2023-07-05
 
-## [0.2.0]
+## 0.2.0
+
 ### Added
+
 - new cipher algorithms
 - remember preferred options (by project)
 
-## [0.1.0] 2022-08-30
+## 0.1.0 2022-08-30
+
 ### Added
+
 - Actions to encrypt or decrypt selected string.
 - the popup asks for password, algorithm and if we want to wrap/unwrap with "ENC()"